# Construct directly the AttributeMap in non-test code.