Benchmarks Before:
```
BenchmarkSpansWithAttributesOCToInternal-16    	  398413	      2933 ns/op	    4867 B/op	      15 allocs/op
BenchmarkSpansWithAttributesUnmarshal-16       	   97926	     11981 ns/op	    5018 B/op	     115 allocs/op
```

Benchmarks After:
```
BenchmarkSpansWithAttributesOCToInternal-16    	  467982	      2418 ns/op	    2328 B/op	      30 allocs/op
BenchmarkSpansWithAttributesUnmarshal-16       	   99662	     12013 ns/op	    5018 B/op	     115 allocs/op
```