@bogdandrutu 2020-04-20
@tigrannajaryan another more trivial optimization (which can give us a lot) is just to have an "IncreseCapacity" for the Map to avoid reconstructing the slice every time we add a new element to the underlying slice (even though we allocate the element in the slice).